Revert object file globbing in Makefile
I feel that file globbing is a bad move in a makefile where the dependency list contains the requisite files. If an additional file is globbed it could cause unexpected problems.
Similarly if a file is missing and no longer required the Make rule will fail but the compile could theoretically have occurred.